build(zero): FAP toolchain + shared/core linkage + embedded float refactor
- application.fam (appid subcensuszero, Sub-GHz category) + minimal ViewDispatcher
  skeleton that builds against the pinned release SDK (API 87.1, target f7)
- subcensus_core.c: unity-build TU that #includes shared/core into the FAP, so
  shared/core stays the single source of truth (no copy/drift) and ufbt builds standalone
- shared/core converted double->float + sqrtf/logf/fabsf/floorf: Cortex-M4F is
  single-precision (-fsingle-precision-constant -Werror=double-promotion); also honors
  Zero §5.5 'cheap enough on-device'. sc_util.h hoists sc_iabs32 for the unity build.
- codegen reconciled with clang-format: emits case-per-line + macro wrap; --check now
  compares CONTENT (whitespace/backslash-normalized) so it stays green against
  ufbt-formatted files. ufbt lint clean.
- ziglang (host clang) + ufbt pinned to release channel.

Green: 47 pytest, 9 native core test files, ufbt build+lint, codegen --check.

+++ shared/core/sc_cadence.h
@@ -25,9 +25,9 @@
 
 typedef struct {
     ScCadenceClass cls;
-    double period_s;    /* estimated fundamental; 0 for event-driven/seen-once (= null) */
-    double regularity;  /* 0..1, 1 = metronomic; 1 - min(1, CoV) (System §7a) */
-    int32_t samples;    /* number of inter-arrival intervals behind the estimate */
+    float period_s; /* estimated fundamental; 0 for event-driven/seen-once (= null) */
+    float regularity; /* 0..1, 1 = metronomic; 1 - min(1, CoV) (System §7a) */
+    int32_t samples; /* number of inter-arrival intervals behind the estimate */
 } ScCadenceEstimate;
 
 /* Estimate cadence from `n` reception timestamps (seconds; need not be pre-sorted). */
@@ -38,14 +38,14 @@ void sc_cadence_from_timestamps(const int64_t* ts_s, size_t n, ScCadenceEstimate
 typedef struct {
     int64_t last_ts;
     int32_t count;
-    double interval_sum;
-    double interval_sqsum;
+    float interval_sum;
+    float interval_sqsum;
     int32_t n_intervals;
-    double bin_width_s;               /* histogram bin width (seconds) */
+    float bin_width_s; /* histogram bin width (seconds) */
     int32_t hist[SC_CADENCE_HIST_BINS];
 } ScCadenceEstimator;
 
-void sc_cadence_init(ScCadenceEstimator* e, double bin_width_s);
+void sc_cadence_init(ScCadenceEstimator* e, float bin_width_s);
 void sc_cadence_observe(ScCadenceEstimator* e, int64_t ts_s);
 void sc_cadence_estimate(const ScCadenceEstimator* e, ScCadenceEstimate* out);
